requirements.txt
discord.py==1.5.0
discord==1.0.1
python-decouple==3.4
